Move TrustPush and related classes to new trust module
Not all trust related classes moved as part of this PR, only main interfaces
to the rest of the code.
