+++ modules/core/src/main/scala/org/tessellation/http/routes/ClusterRoutes.scala
@@ -5,8 +5,6 @@ import cats.syntax.applicativeError._
 import cats.syntax.flatMap._
 import cats.syntax.functor._
 
-import org.tessellation.domain.cluster.programs.TrustPush
-import org.tessellation.domain.trust.storage.TrustStorage
 import org.tessellation.ext.http4s.refined._
 import org.tessellation.schema.cluster._
 import org.tessellation.schema.peer.JoinRequest
@@ -14,6 +12,8 @@ import org.tessellation.schema.peer.Peer.toP2PContext
 import org.tessellation.schema.trust.InternalTrustUpdateBatch
 import org.tessellation.sdk.domain.cluster.programs.{Joining, PeerDiscovery}
 import org.tessellation.sdk.domain.cluster.storage.ClusterStorage
+import org.tessellation.trust.domain.storage.TrustStorage
+import org.tessellation.trust.programs.TrustPush
 
 import com.comcast.ip4s.Host
 import org.http4s.HttpRoutes
