+++ modules/core/src/main/scala/org/tessellation/modules/Storages.scala
@@ -6,16 +6,16 @@ import cats.syntax.functor._
 
 import org.tessellation.config.types.AppConfig
 import org.tessellation.domain.cluster.storage.AddressStorage
-import org.tessellation.domain.trust.storage.TrustStorage
 import org.tessellation.infrastructure.cluster.storage.AddressStorage
 import org.tessellation.infrastructure.db.Database
-import org.tessellation.infrastructure.trust.storage.TrustStorage
 import org.tessellation.sdk.domain.cluster.storage.{ClusterStorage, SessionStorage}
 import org.tessellation.sdk.domain.gossip.RumorStorage
 import org.tessellation.sdk.domain.node.NodeStorage
 import org.tessellation.sdk.infrastructure.cluster.storage.{ClusterStorage, SessionStorage}
 import org.tessellation.sdk.infrastructure.gossip.RumorStorage
 import org.tessellation.sdk.infrastructure.node.NodeStorage
+import org.tessellation.trust.domain.storage.TrustStorage
+import org.tessellation.trust.infrastructure.storage.TrustStorage
 
 object Storages {
